refactor: 딜러와 플레이어가 추상메서드를 확장하도록 구조변경
